update: 添加问题“3315.构造最小位运算数组II”的代码和题解 (#1339)
