--- .commitMsg
@@ -1,8 +1,0 @@
-3314: AC.cpp+py+go+java+rust (#1336) + word(en)
-
-cpp - AC,29.55%,88.64%
-py - AC,31.11%,6.67%
-go - AC,100.00%,33.33%
-java - AC,47.83%,60.87%
-rust - AC,100.00%,80.00%
-rust(inside) - AC,20.00%,80.00%
