+++ Codes/3315-construct-the-minimum-bitwise-array-ii.go
@@ -1,0 +1,5 @@
+package main
+
+func minBitwiseArray(nums []int) []int {
+    
+}
