+++ Codes/3315-construct-the-minimum-bitwise-array-ii.java
@@ -1,0 +1,5 @@
+class Solution {
+    public int[] minBitwiseArray(List<Integer> nums) {
+        
+    }
+}
