+++ Codes/3315-construct-the-minimum-bitwise-array-ii.py
@@ -1,0 +1,5 @@
+from typing import List
+
+class Solution:
+    def minBitwiseArray(self, nums: List[int]) -> List[int]:
+
