+++ Codes/3315-construct-the-minimum-bitwise-array-ii.rs
@@ -1,0 +1,5 @@
+impl Solution {
+    pub fn min_bitwise_array(nums: Vec<i32>) -> Vec<i32> {
+        
+    }
+}
