+++ README.md
@@ -1061,6 +1061,7 @@
 |3305.元音辅音字符串计数I|中等|<a href="https://leetcode.cn/problems/count-of-substrings-containing-every-vowel-and-k-consonants-i/" target="_blank">题目地址</a>|<a href="https://blog.letmefly.xyz/2025/03/12/LeetCode%203305.%E5%85%83%E9%9F%B3%E8%BE%85%E9%9F%B3%E5%AD%97%E7%AC%A6%E4%B8%B2%E8%AE%A1%E6%95%B0I/" target="_blank">题解地址</a>|<a href="https://letmefly.blog.csdn.net/article/details/146197747" target="_blank">CSDN题解</a>|<a href="https://leetcode.cn/problems/count-of-substrings-containing-every-vowel-and-k-consonants-i/solutions/3607571/letmefly-3305yuan-yin-fu-yin-zi-fu-chuan-ptkg/" target="_blank">LeetCode题解</a>|
 |3306.元音辅音字符串计数II|中等|<a href="https://leetcode.cn/problems/count-of-substrings-containing-every-vowel-and-k-consonants-ii/" target="_blank">题目地址</a>|<a href="https://blog.letmefly.xyz/2025/03/13/LeetCode%203306.%E5%85%83%E9%9F%B3%E8%BE%85%E9%9F%B3%E5%AD%97%E7%AC%A6%E4%B8%B2%E8%AE%A1%E6%95%B0II/" target="_blank">题解地址</a>|<a href="https://letmefly.blog.csdn.net/article/details/146228751" target="_blank">CSDN题解</a>|<a href="https://leetcode.cn/problems/count-of-substrings-containing-every-vowel-and-k-consonants-ii/solutions/3609545/letmefly-3306yuan-yin-fu-yin-zi-fu-chuan-7y2q/" target="_blank">LeetCode题解</a>|
 |3314.构造最小位运算数组I|简单|<a href="https://leetcode.cn/problems/construct-the-minimum-bitwise-array-i/" target="_blank">题目地址</a>|<a href="https://blog.letmefly.xyz/2026/01/20/LeetCode%203314.%E6%9E%84%E9%80%A0%E6%9C%80%E5%B0%8F%E4%BD%8D%E8%BF%90%E7%AE%97%E6%95%B0%E7%BB%84I/" target="_blank">题解地址</a>|<a href="https://letmefly.blog.csdn.net/article/details/157185873" target="_blank">CSDN题解</a>|<a href="https://leetcode.cn/problems/construct-the-minimum-bitwise-array-i/solutions/3885383/letmefly-3314gou-zao-zui-xiao-wei-yun-su-12ym/" target="_blank">LeetCode题解</a>|
+|3315.构造最小位运算数组II|中等|<a href="https://leetcode.cn/problems/construct-the-minimum-bitwise-array-ii/" target="_blank">题目地址</a>|<a href="https://blog.letmefly.xyz/2026/01/21/LeetCode%203315.%E6%9E%84%E9%80%A0%E6%9C%80%E5%B0%8F%E4%BD%8D%E8%BF%90%E7%AE%97%E6%95%B0%E7%BB%84II/" target="_blank">题解地址</a>|<a href="https://letmefly.blog.csdn.net/article/details/157225057" target="_blank">CSDN题解</a>|<a href="https://leetcode.cn/problems/construct-the-minimum-bitwise-array-ii/solutions/3886235/letmefly-3315gou-zao-zui-xiao-wei-yun-su-hmbw/" target="_blank">LeetCode题解</a>|
 |3335.字符串转换后的长度I|中等|<a href="https://leetcode.cn/problems/total-characters-in-string-after-transformations-i/" target="_blank">题目地址</a>|<a href="https://blog.letmefly.xyz/2025/05/13/LeetCode%203335.%E5%AD%97%E7%AC%A6%E4%B8%B2%E8%BD%AC%E6%8D%A2%E5%90%8E%E7%9A%84%E9%95%BF%E5%BA%A6I/" target="_blank">题解地址</a>|<a href="https://letmefly.blog.csdn.net/article/details/147916305" target="_blank">CSDN题解</a>|<a href="https://leetcode.cn/problems/total-characters-in-string-after-transformations-i/solutions/3675555/letmefly-3335zi-fu-chuan-zhuan-huan-hou-9ck9y/" target="_blank">LeetCode题解</a>|
 |3337.字符串转换后的长度II|困难|<a href="https://leetcode.cn/problems/total-characters-in-string-after-transformations-ii/" target="_blank">题目地址</a>|<a href="https://blog.letmefly.xyz/2025/05/15/LeetCode%203337.%E5%AD%97%E7%AC%A6%E4%B8%B2%E8%BD%AC%E6%8D%A2%E5%90%8E%E7%9A%84%E9%95%BF%E5%BA%A6II/" target="_blank">题解地址</a>|<a href="https://letmefly.blog.csdn.net/article/details/147976391" target="_blank">CSDN题解</a>|<a href="https://leetcode.cn/problems/total-characters-in-string-after-transformations-ii/solutions/3677448/letmefly-3337zi-fu-chuan-zhuan-huan-hou-h56wo/" target="_blank">LeetCode题解</a>|
 |3340.检查平衡字符串|简单|<a href="https://leetcode.cn/problems/check-balanced-string/" target="_blank">题目地址</a>|<a href="https://blog.letmefly.xyz/2025/03/14/LeetCode%203340.%E6%A3%80%E6%9F%A5%E5%B9%B3%E8%A1%A1%E5%AD%97%E7%AC%A6%E4%B8%B2/" target="_blank">题解地址</a>|<a href="https://letmefly.blog.csdn.net/article/details/146249653" target="_blank">CSDN题解</a>|<a href="https://leetcode.cn/problems/check-balanced-string/solutions/3610899/letmefly-3340jian-cha-ping-heng-zi-fu-ch-p8eo/" target="_blank">LeetCode题解</a>|
