+++ Solutions/LeetCode 3314.构造最小位运算数组I.md
@@ -70,7 +70,7 @@ categories: [题解, LeetCode]
     
 ## 解题方法：模拟
 
-> 今天就每个数$O(1)$的话，明天的每日一题就没得做了。
+> 今天就每个数$O(1)$的话，[明天的](https://blog.letmefly.xyz/2026/01/21/LeetCode%203315.%E6%9E%84%E9%80%A0%E6%9C%80%E5%B0%8F%E4%BD%8D%E8%BF%90%E7%AE%97%E6%95%B0%E7%BB%84II/)每日一题就没得做了。
 
 对于一个数$n$，如何求得其对应的最小$ans$？
