+++ Solutions/Other-English-LearningNotes-SomeWords.md
@@ -1753,6 +1753,9 @@ categories: [自用]
 |Protestant|n. 新教教徒|
 |pilferage|n. 小偷小摸|
 |trainee|n. 实习生|
+|||
+|decree|n. 法令，政令，裁定，判决<br/>v. 裁定，判决，颁布|
+|baggy|adj. 宽松的|
 
 + 这个web要是能设计得可以闭眼(完全不睁眼)键盘控制背单词就好了。
 + 也许可以加个AI用最近词编故事功能(返回接口中支持标注所使用单词高亮?)
